+++ packages/template-retail-react-app/app/components/_app/index.test.js
@@ -19,6 +19,9 @@ import mockConfig from '@salesforce/retail-react-app/config/mocks/default'
 import * as constants from '@salesforce/retail-react-app/app/constants'
 
 jest.mock('../../hooks/use-multi-site', () => jest.fn())
+jest.mock('../../hooks/use-update-shopper-context', () => ({
+    useUpdateShopperContext: jest.fn()
+}))
 
 let windowSpy
 let originalValue