+++ packages/template-retail-react-app/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## v4.1.0-dev (Aug 8, 2024)
 
+### New Features
+- Implement ability to set Shopper Context via search parameters in the Retail React App [#1986](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1986)
+
 ### Performance Improvements
 -   Remove ocapi session-bridging on phased launches [#2011](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2011)